chore: extract override handling into separate file and remove func args
